backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
rasterio==1.4.3
numpy==2.1.3
Pillow==11.0.0
matplotlib==3.9.3
google-generativeai>=0.3.0 
redis>=5.0.0
scikit-learn>=1.3.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
tenacity>=8.2.0
